fixed lseek

+++ kern/include/syscall.h
@@ -61,7 +61,7 @@ int sys___time(userptr_t user_seconds, userptr_t user_nanoseconds);
 int sys__open(const char *filename, int flags, int *retval);
 int sys__read(int fd, void *buf, size_t buflen, int *retval);
 int sys__write(int fd, void *buf, size_t buflen, int *retval);
-int sys__lseek(int fd, off_t pos, int whence, int *retval);
+int sys__lseek(int fd, off_t pos, int whence, int64_t *retval);
 int sys__close(int fd);
 int sys__chdir(const char *pathname);
 int sys__dup2(int oldfd, int newfd, int *retval);
