+++ kern/syscall/lseek.c
@@ -11,7 +11,7 @@
 #include <vnode.h>
 
 int 
-sys__lseek(int fd, off_t pos, int whence, int *retval)
+sys__lseek(int fd, off_t pos, int whence, int64_t *retval)
 {
     lock_acquire(curproc->file_descriptor_table->fd_table_lock);
     if (curproc->file_descriptor_table->count[fd] != 1 || fd >= OPEN_MAX || fd < 0) {
